+++ aldrin/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `UnknownCall`, which represents an unknown pending call.
 - Add `UnknownEvent`, which represents an unknown event.
+- Add variant `Error::InvalidEvent` and inner type `InvalidEvent`.
 
 ### Changed